fix(drive): keep tools.py under the 600-line cap and address review
Split the shared helper layer out of terrarium/drive/tools.py into a new
tools_common module so the file fits the enforced 600-line limit (CI
test-format-guards was failing at 699 lines), and apply the two Copilot
review findings:

- DriveTransitionTool status enum now includes "draft" (it was a
  self-service transition target but absent from the schema enum and the
  unknown-status error message), with a regression test pinning the enum.
- DriveCreateTool docs no longer claim autonomy='continue_when_ready' is
  mandatory: goal autonomy defaults to 'manual' and is only required to
  be a valid mode; the description and schema now say so.
